🔨Added the first frontend implementation test

+++ NearCompanion/Shared/BlockModel.cs
@@ -7,7 +7,7 @@
         public ulong Height { get; set; }
         public ulong LengthMs { get; set; } = 1000;
         public ulong GasPrice { get; set; }
-        public double UtilizationPercentage { get; set; }
+        public decimal UtilizationPercentage { get; set; }
         public ulong TimestampMs { get; set; }
 
         public List<ChunkModel> Chunks { get; set; } = new List<ChunkModel>();
@@ -16,6 +16,6 @@
     public class ChunkModel
     {
         public uint ShardId { get; set; }
-        public double UtilizationPercentage { get; set; }
+        public decimal UtilizationPercentage { get; set; }
     }
 }
